Corrected data type from integers to float on the ingredients_list model by deleting the schema and re-migrating.

+++ recipe_manager_api/db/migrate/20250720020353_create_ingredient_lists.rb
@@ -3,9 +3,9 @@ class CreateIngredientLists < ActiveRecord::Migration[8.0]
     create_table :ingredient_lists do |t|
       t.references :ingredient, null: false, foreign_key: true
       t.references :recipe, null: false, foreign_key: true
-      t.integer :metric_qty
+      t.float :metric_qty
       t.string :metric_unit
-      t.integer :imperial_qty
+      t.float :imperial_qty
       t.string :imperial_unit
 
       t.timestamps
